requirements.txt
cryptography==41.0.4
pycryptodome==3.21.0


